requirements/base.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2==2.9.9
python-environ==0.4.54
